refactor: notification web 객체 Query, Command로 분리
