+++ src/main/java/dandi/dandi/notification/adapter/in/web/NotificationController.java
@@ -2,11 +2,7 @@ package dandi.dandi.notification.adapter.in.web;
 
 import dandi.dandi.auth.adapter.in.web.support.Login;
 import dandi.dandi.notification.application.port.in.NotificationCommandServicePort;
-import dandi.dandi.notification.application.port.in.NotificationQueryServicePort;
-import dandi.dandi.notification.application.port.in.NotificationResponses;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -15,17 +11,9 @@ import org.springframework.web.bind.annotation.RestController;
 public class NotificationController implements NotificationControllerDocs {
 
     private final NotificationCommandServicePort notificationCommandServicePort;
-    private final NotificationQueryServicePort notificationQueryServicePort;
 
-    public NotificationController(NotificationCommandServicePort notificationCommandServicePort,
-                                  NotificationQueryServicePort notificationQueryServicePort) {
+    public NotificationController(NotificationCommandServicePort notificationCommandServicePort) {
         this.notificationCommandServicePort = notificationCommandServicePort;
-        this.notificationQueryServicePort = notificationQueryServicePort;
-    }
-
-    @GetMapping("/notifications")
-    public ResponseEntity<NotificationResponses> getNotifications(@Login Long memberId, Pageable pageable) {
-        return ResponseEntity.ok(notificationQueryServicePort.getNotifications(memberId, pageable));
     }
 
     @PutMapping("/notifications/{notificationId}/check")
